models/nlp/qa/huggingface/requirements.txt
transformers
datasets
torch==1.13.1
valohai-utils~=0.3.1
accelerate
nvidia-ml-py3
